feat: re-enable poison microblock processing, working builds for blockstack_cli, clarity_cli, main
