+++ src/blockstack_cli.rs
@@ -32,6 +32,7 @@ use blockstack_lib::burnchains::bitcoin::address::{
     ADDRESS_VERSION_MAINNET_SINGLESIG, ADDRESS_VERSION_TESTNET_SINGLESIG,
 };
 use blockstack_lib::burnchains::Address;
+use blockstack_lib::chainstate::stacks::StacksBlockHeader;
 use blockstack_lib::chainstate::stacks::{
     StacksBlock, StacksMicroblock, StacksPrivateKey, StacksPublicKey, StacksTransaction,
     StacksTransactionSigner, TokenTransferMemo, TransactionAnchorMode, TransactionAuth,
@@ -43,10 +44,8 @@ use blockstack_lib::codec::{Error as CodecError, StacksMessageCodec};
 use blockstack_lib::core::{CHAIN_ID_MAINNET, CHAIN_ID_TESTNET};
 use blockstack_lib::net::Error as NetError;
 use blockstack_lib::types::chainstate::StacksAddress;
-use blockstack_lib::types::chainstate::StacksBlockHeader;
-use blockstack_lib::util::{
-    hash::hex_bytes, hash::to_hex, log, retry::LogReader, strings::StacksString,
-};
+use blockstack_lib::util::{hash::hex_bytes, hash::to_hex, log, retry::LogReader};
+use blockstack_lib::util_lib::strings::StacksString;
 use blockstack_lib::vm::{
     errors::{Error as ClarityError, RuntimeErrorType},
     types::PrincipalData,
@@ -455,7 +454,7 @@ fn handle_contract_call(
                 Value::try_deserialize_hex_untyped(input)?
             },
             "-e" => {
-                blockstack_lib::clarity::vm_execute(input)?
+                blockstack_lib::clarity_cli::vm_execute(input, &clarity_version)?
                     .ok_or("Supplied argument did not evaluate to a Value")?
             },
             _ => {
