+++ src/chainstate/burn/operations/leader_block_commit.rs
@@ -2265,7 +2265,10 @@ mod tests {
         let mut rng = rand::thread_rng();
         let mut buf = [0u8; 32];
         rng.fill_bytes(&mut buf);
-        let db_path_dir = format!("/tmp/test-blockstack-sortdb-{}", to_hex(&buf));
+        let db_path_dir = format!(
+            "/tmp/stacks-node-tests/unit-tests-sortdb/db-{}",
+            to_hex(&buf)
+        );
 
         let mut db = SortitionDB::connect(
             &db_path_dir,
